Use the linker to resolve classes and support type-safe casting
The Objective-C class `Foo` has a symbol named `_OBJC_CLASS_$_Foo`.
Storing a reference to it in a symbol named
`_OBJC_CLASSLIST_REFERENCES_$_XXX` and then loading it through that
symbol allows the linker to refer to the class directly instead of
having to go through the Objective-C runtime.

Ideally we could use volatile loads to load this symbol, but my testing
shows Zig generates weird (incompatible, and possibly semantically
incorrect) asm when loading a volatile pointer to an extern. So this
uses inline assembly to force the load.

This also adds a helper function named `as()`. It's a way to safely cast
an Objective-C object pointer to a base class or a protocol it
implements. If you try to cast to an unsupported protocol or incorrect
base class it will result in a compile-time error.